+++ src/compose.fxml
@@ -21,31 +21,31 @@
          <tooltip>
             <Tooltip text="Separate receipients with comma" />
          </tooltip></TextField>
-      <Label layoutX="18.0" layoutY="107.0" prefHeight="15.0" prefWidth="29.0" styleClass="description-label" text="CC:" translateX="2.0" />
+      <TextField fx:id="subject" layoutX="76.0" layoutY="70.0" prefHeight="25.0" prefWidth="309.0" promptText="Subject Line" styleClass="input-field">
+         <tooltip>
+            <Tooltip text="Email Subject" />
+         </tooltip></TextField>
       <TextField fx:id="carboncopy" layoutX="49.0" layoutY="102.0" prefHeight="25.0" prefWidth="117.0" promptText="Carbon Copy" styleClass="input-field" />
-      <Label layoutX="178.0" layoutY="107.0" prefHeight="15.0" prefWidth="29.0" styleClass="description-label" text="BCC:" />
       <TextField fx:id="blindcc" layoutX="216.0" layoutY="102.0" prefHeight="25.0" prefWidth="168.0" promptText="Blind Carbon Copy" styleClass="input-field" />
-      <Label layoutX="147.0" layoutY="144.0" prefHeight="15.0" prefWidth="92.0" styleClass="description-label" text="Email content:" />
+      <TextArea fx:id="emailContent" layoutX="35.0" layoutY="161.0" prefHeight="238.0" prefWidth="329.0" styleClass="input-field">
+         <tooltip>
+            <Tooltip text="HTML supported" />
+         </tooltip></TextArea>
       <Button fx:id="attachButton" layoutX="49.0" layoutY="405.0" mnemonicParsing="false" onAction="#addFiles" styleClass="button-login" text="Attach Files">
          <tooltip>
             <Tooltip text="Add files to your email" />
+         </tooltip></Button>
+      <Button layoutX="157.0" layoutY="405.0" mnemonicParsing="false" onAction="#goBack" prefHeight="26.0" prefWidth="72.0" styleClass="button-discard" text="Discard">
+         <tooltip>
+            <Tooltip text="Discard Draft and leave compose screen" />
          </tooltip></Button>
       <Button fx:id="sendButton" layoutX="243.0" layoutY="405.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="109.0" styleClass="button-login" text="Send">
          <tooltip>
             <Tooltip onCloseRequest="#sendEmail" text="Send Email (Unreversable action)" />
          </tooltip></Button>
-      <Label layoutX="21.0" layoutY="75.0" prefHeight="15.0" prefWidth="57.0" styleClass="description-label" text="Subject:" />
-      <TextField fx:id="subject" layoutX="76.0" layoutY="70.0" prefHeight="25.0" prefWidth="309.0" promptText="Subject Line" styleClass="input-field">
-         <tooltip>
-            <Tooltip text="Email Subject" />
-         </tooltip></TextField>
-      <TextArea layoutX="35.0" layoutY="161.0" prefHeight="238.0" prefWidth="329.0" styleClass="input-field">
-         <tooltip>
-            <Tooltip text="HTML supported" />
-         </tooltip></TextArea>
-      <Button layoutX="157.0" layoutY="405.0" mnemonicParsing="false" onAction="#goBack" prefHeight="26.0" prefWidth="72.0" styleClass="button-discard" text="Discard">
-         <tooltip>
-            <Tooltip text="Discard Draft and leave compose screen" />
-         </tooltip></Button>
+      <Label focusTraversable="false" layoutX="18.0" layoutY="107.0" prefHeight="15.0" prefWidth="29.0" styleClass="description-label" text="CC:" translateX="2.0" />
+      <Label focusTraversable="false" layoutX="178.0" layoutY="107.0" prefHeight="15.0" prefWidth="29.0" styleClass="description-label" text="BCC:" />
+      <Label focusTraversable="false" layoutX="147.0" layoutY="144.0" prefHeight="15.0" prefWidth="92.0" styleClass="description-label" text="Email content:" />
+      <Label focusTraversable="false" layoutX="21.0" layoutY="75.0" prefHeight="15.0" prefWidth="57.0" styleClass="description-label" text="Subject:" />
     </children>
 </AnchorPane>
